+++ Assets/Resources/NumCard(dummy).prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 55699752898697921}
   - component: {fileID: 8554550634488803939}
   - component: {fileID: 537758007483561384}
-  - component: {fileID: 6497013270181223760}
+  - component: {fileID: 831210789773202877}
   m_Layer: 5
   m_Name: NumCard(dummy)
   m_TagString: NumCrad
@@ -140,9 +140,9 @@ Rigidbody2D:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5616394500776527374}
-  m_BodyType: 0
+  m_BodyType: 1
   m_Simulated: 1
-  m_UseFullKinematicContacts: 0
+  m_UseFullKinematicContacts: 1
   m_UseAutoMass: 0
   m_Mass: 1
   m_LinearDrag: 0
@@ -159,7 +159,7 @@ Rigidbody2D:
   m_SleepingMode: 0
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!114 &6497013270181223760
+--- !u!114 &831210789773202877
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
