+++ ProjectSettings/CommonBurstAotSettings.json
@@ -1,0 +1,6 @@
+{
+  "MonoBehaviour": {
+    "Version": 4,
+    "DisabledWarnings": ""
+  }
+}
