+++ ProjectSettings/EditorBuildSettings.asset
@@ -8,10 +8,13 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/TopScene.unity
     guid: 3e8614603f3fcdd44af3b40ce67e4ad5
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/Ranking.unity
     guid: 2effdf7f69747a642a18e3dd384d4f86
+  - enabled: 1
+    path: Assets/Scenes/GameScene.unity
+    guid: 59349abf7ea0db149bdfa9301ee02456
   m_configObjects: {}
